Add comments and reorder methods
