+++ src/main/java/frc/robot/subsystems/shooter/Shooter.java
@@ -8,12 +8,15 @@ public class Shooter extends SpikeSystem<ShooterIO.ShooterIOInputs> {
     private static final double SHOOTER_READY_THRESHOLD_RPS = 0.5; // RPS threshold to consider the shooter ready
 
     private ShooterIO shooterIO;
-    private double targetRPS = 0.0;
+    private double targetRPS = 0.0; // Target rotations per second
 
     public Shooter() {
         super("Shooter", new ShooterIO.ShooterIOInputs());
     }
 
+    /**
+     *  Set rps and hood angle from adjusted shot data
+     */ 
     @Override
     public void onPeriodic() {
         ShotCompensation.AdjustedShot shotData = Targeting.getShotData();
@@ -27,12 +30,19 @@ public class Shooter extends SpikeSystem<ShooterIO.ShooterIOInputs> {
         }
     }
 
+    /**
+     * Sets up the data refresher for Shooter
+     */
     @Override
     protected Runnable setupDataRefresher() {
         shooterIO = new ShooterIOTalonFX();
         return useAsyncDataRefresher(shooterIO);
     }
 
+    /**
+     * Checks if current rps of the motor is within the allowed error bounds 
+     * @return if the target is within error bounds 
+     */
     public boolean isAtTargetRPS() {
         return Math.abs(super.io.motorRPS - targetRPS) < SHOOTER_READY_THRESHOLD_RPS;
     }
